+++ custom_components/keymaster/helpers.py
@@ -1,4 +1,5 @@
 """Helpers for keymaster."""
+from datetime import timedelta
 import logging
 import os
 from typing import Dict, List, Optional, Union
@@ -12,11 +13,27 @@ from homeassistant.components.ozw import DOMAIN as OZW_DOMAIN
 from homeassistant.components.timer import DOMAIN as TIMER_DOMAIN
 from homeassistant.components.zwave.const import DOMAIN as ZWAVE_DOMAIN
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant
+from homeassistant.const import STATE_LOCKED, STATE_UNLOCKED
+from homeassistant.core import HomeAssistant, State
 from homeassistant.helpers.entity_registry import async_get_registry
+from homeassistant.util import datetime as dt_util
 from homeassistant.util.yaml.loader import load_yaml
 
-from .const import ATTR_NODE_ID, CONF_PATH, DOMAIN, PRIMARY_LOCK
+from .const import (
+    ACCESS_CONTROL,
+    ACTION_MAP,
+    ALARM_TYPE,
+    ATTR_ACTION_CODE,
+    ATTR_ACTION_TEXT, ATTR_NAME,
+    ATTR_NODE_ID,
+    ATTR_USER_CODE,
+    ATTR_USER_CODE_NAME,
+    CONF_PATH,
+    DOMAIN,
+    EVENT_KEYMASTER_LOCK_STATE_CHANGED,
+    LOCK_STATE_MAP,
+    PRIMARY_LOCK,
+)
 from .lock import KeymasterLock
 
 _LOGGER = logging.getLogger(__name__)
@@ -140,3 +157,79 @@ def delete_folder(absolute_path: str, *relative_paths: str) -> None:
         for file_or_dir in os.listdir(path):
             delete_folder(path, file_or_dir)
         os.rmdir(path)
+
+
+def handle_state_change(
+    hass: HomeAssistant,
+    config_entry: ConfigEntry,
+    changed_entity: str,
+    new_state: State,
+) -> None:
+    """Listener to track state changes to lock entities."""
+    primary_lock: KeymasterLock = hass.data[DOMAIN][config_entry.entry_id][PRIMARY_LOCK]
+
+    # If listener was called for entity that is not for this entry, ignore
+    if changed_entity not in [
+        primary_lock.lock_entity_id,
+        primary_lock.alarm_level_or_user_code_entity_id,
+        primary_lock.alarm_type_or_access_control_entity_id,
+    ]:
+        return
+
+    # Determine action type to set appropriate action text using ACTION_MAP
+    action_type = ""
+    if ALARM_TYPE in primary_lock.alarm_type_or_access_control_entity_id:
+        action_type = ALARM_TYPE
+    if ACCESS_CONTROL in primary_lock.alarm_type_or_access_control_entity_id:
+        action_type = ACCESS_CONTROL
+
+    # Get alarm_level/usercode and alarm_type/access_control  states
+    alarm_level_state = hass.states.get(primary_lock.alarm_level_or_user_code_entity_id)
+    alarm_level_value = int(alarm_level_state.state) if alarm_level_state else None
+
+    alarm_type_state = hass.states.get(
+        primary_lock.alarm_type_or_access_control_entity_id
+    )
+    alarm_type_value = int(alarm_type_state.state) if alarm_type_state else None
+
+    # If lock has changed state but alarm_type/access_control state hasn't changed in a while
+    # set action_value to RF lock/unlock
+    if changed_entity == primary_lock.lock_entity_id:
+        if (
+            alarm_level_state is None
+            or int(alarm_level_state.state) != 0
+            or (dt_util.utcnow() - alarm_type_state.last_changed < timedelta(seconds=2))
+        ):
+            return
+
+        if (
+            new_state.state in (STATE_LOCKED, STATE_UNLOCKED)
+            and action_type in LOCK_STATE_MAP
+        ):
+            alarm_type_value = LOCK_STATE_MAP[action_type][new_state.state]
+
+    action_text = (
+        ACTION_MAP.get(action_type, {}).get(
+            alarm_type_value, "Unknown Alarm Type Value"
+        )
+        if alarm_type_value is not None
+        else None
+    )
+    usercode_name = (
+        hass.states.get(
+            f"input_text.{primary_lock.lock_name}_name_{alarm_level_value}"
+        ).state
+        if alarm_level_value is not None
+        else None
+    )
+
+    hass.bus.async_fire(
+        EVENT_KEYMASTER_LOCK_STATE_CHANGED,
+        event_data={
+            ATTR_NAME: primary_lock.lock_name,
+            ATTR_ACTION_CODE: alarm_type_value,
+            ATTR_ACTION_TEXT: action_text,
+            ATTR_USER_CODE: alarm_level_value,
+            ATTR_USER_CODE_NAME: usercode_name,
+        },
+    )
